msm: camera: reqmgr: save frame id from trigger
Current logic for dual cam use-case skips frame when difference
between SOF counters in both link and sync link for that frame
is more than 1.

The patch modifies this behavior so then frame id associated with SOF
received for particular link is compared with frame id associated
with SOF received for sync link and frame is skipped only when these
frame ids are also different.

Change-Id: I9138352d5849c645e057ba33e25a268efdc061a3
Signed-off-by: Ivan Vecera <ivan@cera.cz>